payment-service/requirements.txt
Flask==2.3.3
pytest==8.4.2
pyyaml==6.0.2
openapi-core==0.18.1